# src/main/resources/static/hackhub-openapi.yaml
openapi: 3.0.0
info:
  title: hackhub-openapi
  version: 1.0.0
paths:
...
  /api/teams:
    post:
      summary: Crea un nuovo team
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_team_crea_un_nuovo_team_yml
      description: ''
      tags:
        - Team
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  nome:
                    type: string
                  maxMembri:
                    type: integer
                  membri:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        nome:
                          type: string
                        cognome:
                          type: string
                        role:
                          type: string
              example:
                nome: ''
                maxMembri: 0
                membri:
                  - id: ''
                    nome: ''
                    cognome: ''
                    role: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      requestBody:
        $ref: '#/components/requestBodies/crea_un_nuovo_team'
    get:
      summary: Restituisce tutti i team (opzionalmente filtrati per hackathon)
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_team_restituisce_tutti_i_team_opzionalmente_filtrati_per_hackathon_yml
      description: ''
      tags:
        - Team
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    nome:
                      type: string
                    maxMembri:
                      type: integer
                    membri:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                          nome:
                            type: string
                          cognome:
                            type: string
                          role:
                            type: string
              example:
                - nome: ''
                  maxMembri: 0
                  membri:
                    - id: ''
                      nome: ''
                      cognome: ''
                      role: ''
      parameters:
        - name: hackathonId
          in: query
          description: ''
          required: false
          schema:
            type: string
  /api/teams/:nome/iscrizione?hackathonId=e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3:
    post:
      summary: Iscrive un team a un hackathon
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_team_iscrive_un_team_a_un_hackathon_yml
      description: ''
      tags:
        - Team
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: hackathonId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: e8f9b1c2-8f80-4ef1-95df-462f3b4f11f3
        - name: nome
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: team fortissimo
  /api/teams/esci?utenteId=94fca499-7f34-4f41-bc0a-82684187a6b3:
    post:
      summary: Permette a un utente di uscire dal proprio team
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_team_permette_a_un_utente_di_uscire_dal_proprio_team_yml
      description: ''
      tags:
        - Team
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: utenteId
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: 94fca499-7f34-4f41-bc0a-82684187a6b3
  /api/teams/:nome:
    get:
      summary: Restituisce un team specifico
      operationId: >-
        c_users_heron_onedrive_documenti_bruno_hackhub_api_-_2_team_restituisce_un_team_specifico_yml
      description: ''
      tags:
        - Team
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  nome:
                    type: string
                  maxMembri:
                    type: integer
                  membri:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        nome:
                          type: string
                        cognome:
                          type: string
                        role:
                          type: string
              example:
                nome: ''
                maxMembri: 0
                membri:
                  - id: ''
                    nome: ''
                    cognome: ''
                    role: ''
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: ''
      parameters:
        - name: nome
          in: path
          description: ''
          required: true
          schema:
            type: string
          example: piedi
...
servers:
  - url: http://localhost:8080
    description: Server di sviluppo
components:
  schemas:
...
    crea_un_nuovo_team:
      type: object
      properties:
        utenteId:
          type: string
        nome:
          type: string
        maxMembri:
          type: integer
      example:
        utenteId: 690f6c12-4331-42f9-ab66-258bc93d1fed
        nome: team fortissimo
        maxMembri: 3
...
  requestBodies:
...
    crea_un_nuovo_team:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/crea_un_nuovo_team'
      description: ''
      required: true
...
  securitySchemes: {}